Validate user status method and update regEx email
There was a bug in the email field if the user typed an email with a "ñ"
word, that's why the regular expression was update.
The user status now can accept a null value because it is not required,
and also we validate that the status won't be a value between 0 or 1.
When we validate the first name and the last name i notice that the user
can type in capital letters and small letters